+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/roadrunner/SampleMecanumDrive.java
@@ -28,6 +28,7 @@ import com.qualcomm.robotcore.hardware.VoltageSensor;
 import com.qualcomm.robotcore.hardware.configuration.typecontainers.MotorConfigurationType;
 
 import org.firstinspires.ftc.robotcore.external.navigation.AngleUnit;
+import org.firstinspires.ftc.teamcode.modules.drivetrainrr.DriveConstants;
 import org.firstinspires.ftc.teamcode.roadrunner.trajectorysequence.TrajectorySequence;
 import org.firstinspires.ftc.teamcode.roadrunner.trajectorysequence.TrajectorySequenceBuilder;
 import org.firstinspires.ftc.teamcode.roadrunner.trajectorysequence.TrajectorySequenceRunner;
@@ -37,10 +38,10 @@ import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 
-import static org.firstinspires.ftc.teamcode.roadrunner.DriveConstants.*;
-import static org.firstinspires.ftc.teamcode.roadrunner.DriveConstants.MAX_ACCEL;
-import static org.firstinspires.ftc.teamcode.roadrunner.DriveConstants.MAX_ANG_VEL;
-import static org.firstinspires.ftc.teamcode.roadrunner.DriveConstants.MAX_VEL;
+import static org.firstinspires.ftc.teamcode.modules.drivetrainrr.DriveConstants.*;
+import static org.firstinspires.ftc.teamcode.modules.drivetrainrr.DriveConstants.MAX_ACCEL;
+import static org.firstinspires.ftc.teamcode.modules.drivetrainrr.DriveConstants.MAX_ANG_VEL;
+import static org.firstinspires.ftc.teamcode.modules.drivetrainrr.DriveConstants.MAX_VEL;
 
 /*
  * Simple mecanum drive hardware implementation for REV hardware.
